[REF] cookingo_custom: Make testcase common
